+++ lib/components/IconButton/IconButton.d.ts
@@ -4,8 +4,9 @@ export interface IIconButtonProps {
     icon?: string;
     color?: COLOR;
     disabled?: boolean;
+    isActive?: boolean;
     onClick?: (e: MouseEvent) => void;
     variant?: VARIANT;
     className?: string;
 }
-export declare const IconButton: ({ icon, color, disabled, onClick, variant, className }: IIconButtonProps) => JSX.Element;
+export declare const IconButton: ({ icon, color, disabled, isActive, onClick, variant, className }: IIconButtonProps) => JSX.Element;