+++ src/lib/components/Drawer/Drawer.tsx
@@ -1,14 +1,22 @@
-import React, { ReactNode } from 'react';
+import React, { ReactNode, useEffect } from 'react';
 import { createPortal } from 'react-dom';
 import { Backdrop } from '../Backdrop';
 
 export interface IDrawerProps {
 	children?: ReactNode;
 	position?: 'left' | 'right';
+	// closeOnBackdropClick?: boolean;
 	onClickBackdrop?: () => void;
 }
 
 export const Drawer = ({ children, position = 'left', onClickBackdrop }: IDrawerProps) => {
+
+	useEffect(() => {
+		document.body.classList.add('modal-open');
+		return () => {
+			document.body.classList.remove('modal-open');
+		};
+	}, []);
 
 	const handleClickBackdrop = () => {
 		onClickBackdrop && onClickBackdrop();