+++ src/pages/Showcase/assets/icons.ts
@@ -1,5 +1,7 @@
 import infoSolidSvg from './info-solid.svg';
+import barsSolidSvg from './bars-solid.svg';
 
 export {
-	infoSolidSvg
+	infoSolidSvg,
+	barsSolidSvg
 };